Struct cv_core::WorldToCamera

pub struct WorldToCamera(pub IsometryMatrix3<f64>);

This contains a world pose, which is a pose of the world relative to the camera. This maps WorldPoint into CameraPoint, changing an absolute position into a vector relative to the camera.

Tuple Fields
0: IsometryMatrix3<f64>