Trait cv_core::CameraModel

pub trait CameraModel {
    type Projection: Bearing;
    fn calibrate<P>(&self, point: P) -> Self::Projection
    where
        P: ImagePoint;

    fn uncalibrate(&self, projection: Self::Projection) -> KeyPoint;
}

Allows conversion between the point on an image and the internal projection which can describe the bearing of the projection out of the camera.

Associated Types

type Projection: Bearing

Required methods

fn calibrate<P>(&self, point: P) -> Self::Projection where
P: ImagePoint,

Extracts a projection from a pixel location in an image.

fn uncalibrate(&self, projection: Self::Projection) -> KeyPoint

Extracts the pixel location in the image from the projection.